# server/contracts/openapi.yml
...
openapi: 3.0.3
info:
  title: Notebook Wars metaserver REST API
  version: 0.1.0
  description: |
    Single source of truth for the REST contract (design-first, M15). The metaserver uses fastify-openapi-glue to wire routes + validation by operationId; the client uses openapi-typescript + openapi-fetch to generate typed fetch. Unified response envelope ApiResp<T>. See SERVER_API.md.
servers:
  - url: /api
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AnalyticsEvent:
      type: object
      required:
        - event
        - ts
      properties:
        event:
          type: string
        ts:
          type: number
          description: Client unix ms
        props:
          type: object
          additionalProperties: true
    AnalyticsEventBatch:
      type: object
      required:
        - events
      properties:
        session_id:
          type: string
        device_id:
          type: string
        platform:
          type: string
          enum:
            - web
            - wechat
            - crazygames
        os:
          type: string
        game_version:
          type: string
        locale:
          type: string
        ua:
          type: string
        screen_w:
          type: integer
        screen_h:
          type: integer
        dpr:
          type: number
        consent:
          type: boolean
        events:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/AnalyticsEvent'
...
paths:
...
  /analytics/events:
    post:
      tags:
        - analytics
      operationId: postAnalyticsEvents
      summary: Batch event reporting (JWT optional; attaches user_id when token present, otherwise anonymous)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticsEventBatch'
      responses:
        '200':
          description: Accepted (fire-and-forget, does not guarantee persistence)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    enum:
                      - true
                  data:
                    nullable: true